requirements.txt
# Discord Bot
discord.py>=2.3.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dotenv>=1.0.0
urllib3>=2.0.0

# Database (SQLite is built-in to Python)
# No additional requirements for SQLite

# Web UI Dependencies
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
jinja2>=3.1.0

# Optional: For advanced features
# selenium>=4.15.0  # For JavaScript-heavy sites
# undetected-chromedriver>=3.5.0  # For anti-bot bypass